Optimize logaddexp and friends
