+++ funsor/einsum/numpy_log.py
@@ -24,7 +24,7 @@ def einsum(equation, *operands):
     shifts = []
     exp_operands = []
     for dims, operand in zip(inputs, operands):
-        shift = operand
+        shift = ops.detach(operand)
         for i, dim in enumerate(dims):
             if dim not in output:
                 shift = ops.amax(shift, i, keepdims=True)
